// src/app/login/Model/resposta.ts
// export class valores {
//     email: string | undefined;
//     senha: string | undefined;
//     first_name: string | undefined;
//     last_name: string | undefined;
// }

// export class valoresPost {
//     email: string | undefined;
//     senha: string | undefined;
//     first_name: string | undefined;
//     last_name: string | undefined;
// }

export class valores {
    id: number | undefined;
    email: string | undefined;
    first_name: string | undefined;
    last_name: string | undefined;
}

export class valoresPost {
    email: string | undefined;    //A interrogação indica que esse valor pode existir ou não
    first_name: string | undefined;
    last_name: string | undefined;
    senha: string | undefined;
}